Move Theme to the top of the crate
